update Particle module, replace NucleiProp dict with usage of particle package and NuDat3 table

+++ Particle/Generators/Distributions.py
@@ -172,8 +172,8 @@ class Disk(AbsDistribution):
 class UserInput(AbsDistribution):
     def __init__(self, R0=np.zeros((1, 3)), V0=np.zeros((1, 3)), *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.r = np.array(R0, ndmin=2)
-        self.v = np.array(V0, ndmin=2)
+        self.r = np.array(R0, dtype=np.float64, ndmin=2)
+        self.v = np.array(V0, dtype=np.float64, ndmin=2)
 
     def generate_coordinates(self, *args, **kwargs):
         if self.r.shape != (self.flux.Nevents, 3) or self.v.shape != (self.flux.Nevents, 3):
